Time: 0 ms (100%), Space: 47.5 MB (89.15%) - LeetHub

+++ 0050-powx-n/0050-powx-n.java
@@ -2,6 +2,6 @@ class Solution
 {
     public double myPow(double x, int n) 
     {
-        return  (double)Math.pow(x, n) ;
+        return   Math.pow(x, n) ;
     }
 }
